+++ tests/integ/datasource/test_oracledb.py
@@ -21,7 +21,7 @@ from snowflake.snowpark._internal.data_source.utils import (
 )
 from snowflake.snowpark.types import StructType, StructField, StringType
 from snowflake.snowpark.exceptions import (
-    SnowparkDataSourceNonRetryableException,
+    _SnowparkDataSourceNonRetryableException,
     SnowparkDataframeReaderException,
     SnowparkSQLException,
 )
@@ -257,7 +257,7 @@ def test_unsupported_type():
 
 def test_oracledb_non_retryable_error(session):
     with pytest.raises(
-        SnowparkDataSourceNonRetryableException,
+        _SnowparkDataSourceNonRetryableException,
         match="ORA-00920: invalid relational operator",
     ):
         session.read.dbapi(