+++ CHANGELOG.md
@@ -56,6 +56,9 @@
   - `nvl2`
   - `regr_valx`
 
+#### Improvements
+
+- Improve `DataFrameReader.dbapi`(PuPr) that dbapi will not retry on non-retryable error such as SQL syntax error on external data source query.
 
 ### Snowpark pandas API Updates
 
@@ -67,7 +70,6 @@
 - Hybrid execution mode is now enabled by default. Certain operations on smaller data will now automatically execute in native pandas in-memory. Use `from modin.config import AutoSwitchBackend; AutoSwitchBackend.disable()` to turn this off and force all execution to occur in Snowflake.
 - Added a session parameter `pandas_hybrid_execution_enabled` to enable/disable hybrid execution as an alternative to using `AutoSwitchBackend`.
 - Removed an unnecessary `SHOW OBJECTS` query issued from `read_snowflake` under certain conditions.
-- Add non-retryable failure in `DataFrameReader.dbapi`(PuPr), such as syntax error in external data source SQL.
 
 ## 1.39.0 (2025-09-17)